rename some components names & show list actions only to admin
