+++ src/containers/Invoice/List/List.jsx
@@ -19,6 +19,7 @@ import { ModalInviteContractor } from "src/containers/User/Invite";
 import styles from "./List.module.css";
 import { Row } from "src/components/layout";
 import useModalContext from "src/hooks/utils/useModalContext";
+import AdminContent from "src/components/AdminContent";
 
 const ActionsContent = ({ openInviteModal }) => {
   const mobile = useMediaQuery("(max-width: 768px)");
@@ -45,6 +46,7 @@ const ActionsContent = ({ openInviteModal }) => {
   );
 
   return (
+    <AdminContent>
     <>
       {!mobile ? (
         <Row justify="space-between" className={styles.actionsWrapper}>
