+++ src/containers/Invoice/List/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ListWrapper";
